refactor(store): Move app.actions.ts to shared module
