+++ shared/src/store/actions/app.action.ts
@@ -1,5 +1,5 @@
 import { Action } from '@ngrx/store';
-import { type } from '../../shared/util/';
+import { type } from '../../util';
 
 const PREFIX = '[app] ';
